apps/api/requirements.txt
fastapi>=0.115.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
sqlalchemy[asyncio]>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.10
asyncpg>=0.30.0
PyYAML
python-multipart>=0.0.9
httpx>=0.27.0
anthropic>=0.40.0
boto3>=1.34.0
cloudinary>=1.36.0
Pillow>=10.0.0
